requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
requests==2.31.0
python-dotenv==1.0.1
selenium-wire
selenium==4.18.1
beautifulsoup4==4.12.3
python-dateutil==2.8.2
blinker==1.4.0
zoneinfo==0.2.1; python_version < '3.9'
pydantic==2.6.3

